move training loop for profiler
